requirements.txt
aiohttp==3.6.2
apiai==1.2.3
asgiref==3.2.7
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudipsp==1.0.2
colorama==0.4.3
cycler==0.10.0
discord==1.0.1
discord.py==1.5.0
docopt==0.6.2
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
packaging==20.4
Pillow==7.2.0
pip==20.2.3
pipwin==0.5.0
playsound==1.2.2
pycodestyle==2.6.0
pycparser==2.20
pyjsparser==2.7.1
pylint==2.5.3
PyNaCl==1.3.0
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
setuptools==47.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
yarl==1.6.0
youtube-dl==2020.9.20
zope.event==4.4
zope.interface==5.1.0